backend/requirements/base.txt
# SHA1:feb5ffffea373beacd878345c8e626e300cf7081
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
aiohttp==3.9.1
    # via tuspy
aiosignal==1.3.1
    # via aiohttp
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bioblend==1.2.0
    # via -r backend/requirements/base.in
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==41.0.7
    # via
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
defusedxml==0.7.1
    # via
    #   fpdf2
    #   odfpy
diff-match-patch==20230430
    # via django-import-export
dj-database-url==2.1.0
    # via -r backend/requirements/base.in
django==4.2.9
    # via
    #   -r backend/requirements/base.in
    #   dj-database-url
    #   django-authtools
    #   django-extensions
    #   django-import-export
    #   django-linear-migrations
    #   django-picklefield
    #   djangorestframework
    #   drf-spectacular
    #   mozilla-django-oidc
django-admin-list-filter-dropdown==1.0.3
    # via -r backend/requirements/base.in
django-authtools @ git+https://github.com/adRn-s/django-authtools@dev_adRn
    # via -r backend/requirements/base.in
django-constance[database]==3.1.0
    # via -r backend/requirements/base.in
django-extensions==3.2.3
    # via -r backend/requirements/base.in
django-import-export==3.3.6
    # via -r backend/requirements/base.in
django-linear-migrations==2.12.0
    # via -r backend/requirements/base.in
django-monthfield==0.1.4
    # via -r backend/requirements/base.in
django-picklefield==3.1
    # via django-constance
djangorestframework==3.14.0
    # via
    #   -r backend/requirements/base.in
    #   drf-spectacular
drf-spectacular==0.27.1
    # via -r backend/requirements/base.in
et-xmlfile==1.1.0
    # via openpyxl
fonttools==4.47.2
    # via fpdf2
fpdf2==2.7.7
    # via -r backend/requirements/base.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gunicorn==21.2.0
    # via -r backend/requirements/base.in

idna==3.6
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
josepy==1.14.0
    # via mozilla-django-oidc
jsonschema==4.21.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
lxml==5.1.0
    # via python-docx
markuppy==1.14
    # via tablib
mozilla-django-oidc==3.0.0
    # via -r backend/requirements/base.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.26.3
    # via
    #   -r backend/requirements/base.in
    #   pandas
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==23.2
    # via gunicorn
pandas==2.1.4
    # via -r backend/requirements/base.in
pillow==10.2.0
    # via fpdf2
psycopg2-binary==2.9.9
    # via -r backend/requirements/base.in
pycparser==2.21
    # via cffi
pyopenssl==23.3.0
    # via josepy
python-dateutil==2.8.2
    # via
    #   -r backend/requirements/base.in
    #   pandas
python-docx==1.1.0
    # via -r backend/requirements/base.in
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   pandas
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   tablib
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   bioblend
    #   mozilla-django-oidc
    #   requests-toolbelt
    #   tuspy
requests-toolbelt==1.0.0
    # via bioblend
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via django-import-export
tblib==3.0.0
    # via -r backend/requirements/base.in
tinydb==4.8.0
    # via tuspy
tuspy==1.0.3
    # via bioblend
typing-extensions==4.9.0
    # via
    #   bioblend
    #   dj-database-url
    #   python-docx
tzdata==2023.4
    # via pandas
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.1.0
    # via requests
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via
    #   -r backend/requirements/base.in
    #   tablib
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
setuptools==69.0.3
    # via -r backend/requirements/base.in
